Setup: Deleted alphabet position files and moved extract new time to utils
